replace Int64 with Int to improve compatibility

+++ scripts/converter.jl
@@ -23,7 +23,7 @@ end
 
 accepted_keys = ["junction"]
 for (i, junction) in data["junction"]
-    junction["id"] = (get(junction, "junction_i", false) == true) ? Int(junction["junction_i"]) : parse(Int64, i)
+    junction["id"] = (get(junction, "junction_i", false) == true) ? Int(junction["junction_i"]) : parse(Int, i)
     if (data["per_unit"] == 1)
         junction["p_min"] = junction["pmin"] * data["baseP"]
         junction["p_max"] = junction["pmax"] * data["baseP"]
@@ -51,7 +51,7 @@ end
 
 push!(accepted_keys, "pipe")
 for (i, pipe) in data["pipe"]
-    pipe["id"] = (get(pipe, "pipline_i", false) != false) ? Int(pipe["pipline_i"]) : parse(Int64, i)
+    pipe["id"] = (get(pipe, "pipline_i", false) != false) ? Int(pipe["pipline_i"]) : parse(Int, i)
     pipe["fr_junction"] = pipe["f_junction"]
     pipe["to_junction"] = pipe["t_junction"]
     pipe["p_min"] = min(
@@ -72,7 +72,7 @@ end
 push!(accepted_keys, "compressor")
 for (i, compressor) in data["compressor"]
     compressor["status"] = 1
-    compressor["id"] = (get(compressor, "compressor_i", false) != false) ? Int(compressor["compressor_i"]) : parse(Int64, i)
+    compressor["id"] = (get(compressor, "compressor_i", false) != false) ? Int(compressor["compressor_i"]) : parse(Int, i)
     compressor["fr_junction"] = Int(compressor["f_junction"])
     compressor["to_junction"] = Int(compressor["t_junction"])
     if get(compressor, "qmin", false) == false
@@ -121,7 +121,7 @@ push!(accepted_keys, "receipt")
 data["receipt"] = Dict()
 for (i, producer) in data["producer"]
     data["receipt"][i] = Dict{String,Any}()
-    data["receipt"][i]["id"] = (get(producer, "producer_i", false) != false) ? Int(producer["producer_i"]) : parse(Int64, i)
+    data["receipt"][i]["id"] = (get(producer, "producer_i", false) != false) ? Int(producer["producer_i"]) : parse(Int, i)
     data["receipt"][i]["junction_id"] = (get(producer, "junction", false) != false) ? producer["junction"] : producer["qg_junc"]
     data["receipt"][i]["status"] = (get(producer, "status", false) != false) ? producer["status"] : Int(1)
     data["receipt"][i]["is_dispatchable"] = producer["dispatchable"]
@@ -141,7 +141,7 @@ push!(accepted_keys, "delivery")
 data["delivery"] = Dict()
 for (i, consumer) in data["consumer"]
     data["delivery"][i] = Dict{String,Any}()
-    data["delivery"][i]["id"] = (get(consumer, "consumer_i", false) != false) ? Int(consumer["consumer_i"]) : parse(Int64, i)
+    data["delivery"][i]["id"] = (get(consumer, "consumer_i", false) != false) ? Int(consumer["consumer_i"]) : parse(Int, i)
     data["delivery"][i]["junction_id"] = (get(consumer, "junction", false) != false) ? consumer["junction"] : consumer["ql_junc"]
     data["delivery"][i]["status"] = (get(consumer, "status", false) != false) ? Int(consumer["status"]) : Int(1)
     data["delivery"][i]["is_dispatchable"] = consumer["dispatchable"]
@@ -162,7 +162,7 @@ delete!(data, "consumer")
 
 push!(accepted_keys, "valve")
 for (i, valve) in get(data, "valve", [])
-    valve["id"] = parse(Int64, i)
+    valve["id"] = parse(Int, i)
     valve["fr_junction"] = valve["f_junction"]
     valve["to_junction"] = valve["t_junction"]
     valve["status"] = 1
@@ -172,7 +172,7 @@ end
 
 push!(accepted_keys, "short_pipe")
 for (i, short_pipe) in get(data, "short_pipe", [])
-    short_pipe["id"] = parse(Int64, i)
+    short_pipe["id"] = parse(Int, i)
     short_pipe["fr_junction"] = short_pipe["f_junction"]
     short_pipe["to_junction"] = short_pipe["t_junction"]
     short_pipe["status"] = 1
@@ -183,7 +183,7 @@ end
 
 push!(accepted_keys, "resistor")
 for (i, resistor) in get(data, "resistor", [])
-    resistor["id"] = parse(Int64, i)
+    resistor["id"] = parse(Int, i)
     resistor["fr_junction"] = resistor["f_junction"]
     resistor["to_junction"] = resistor["t_junction"]
     resistor["status"] = 1
@@ -196,7 +196,7 @@ push!(accepted_keys, "regulator")
 data["regulator"] = Dict()
 for (i, control_valve) in get(data, "control_valve", [])
     data["regulator"][i] = Dict{String,Any}()
-    data["regulator"][i]["id"] = parse(Int64, i)
+    data["regulator"][i]["id"] = parse(Int, i)
     data["regulator"][i]["fr_junction"] = control_valve["f_junction"]
     data["regulator"][i]["to_junction"] = control_valve["t_junction"]
     data["regulator"][i]["reduction_factor_min"] = control_valve["c_ratio_min"]
@@ -222,7 +222,7 @@ end
 if haskey(data, "ne_pipe")
     push!(accepted_keys, "ne_pipe")
     for (i, pipe) in data["ne_pipe"]
-        pipe["id"] = (get(pipe, "pipline_i", false) != false) ? Int(pipe["pipline_i"]) : parse(Int64, i)
+        pipe["id"] = (get(pipe, "pipline_i", false) != false) ? Int(pipe["pipline_i"]) : parse(Int, i)
         pipe["fr_junction"] = pipe["f_junction"]
         pipe["to_junction"] = pipe["t_junction"]
         pipe["p_min"] = min(
@@ -245,7 +245,7 @@ if haskey(data, "ne_compressor")
     push!(accepted_keys, "ne_compressor")
     for (i, compressor) in data["ne_compressor"]
         compressor["status"] = 1
-        compressor["id"] = (get(compressor, "compressor_i", false) != false) ? Int(compressor["compressor_i"]) : parse(Int64, i)
+        compressor["id"] = (get(compressor, "compressor_i", false) != false) ? Int(compressor["compressor_i"]) : parse(Int, i)
         compressor["fr_junction"] = Int(compressor["f_junction"])
         compressor["to_junction"] = Int(compressor["t_junction"])
         if get(compressor, "qmin", false) == false
